Schedule Workout Fixed Schedule Workout Modal moved to Child Component. Calling from Parent was displaying all modals at once
